feat: float16 load

+++ extra/datasets/imagenet.py
@@ -123,7 +123,7 @@ def iterate(bs=32, val=True, shuffle=True, num_workers=0):
       X = [image_load(files[i], val) for i in order[i:i+bs]]
     Y = [cir[files[i].split("/")[-2]] for i in order[i:i+bs]]
     et = time.monotonic()
-    yield (np.array(X), np.array(Y), et - st)
+    yield (np.array(X, dtype=np.float16), np.array(Y, dtype=np.float16), et - st)
 
 def fetch_batch(bs, val=False):
   files = get_val_files() if val else get_train_files()
